General cleanup (#11)
The module and commonjs builds no longer bundle mapbox-style-spec. This is good, since it's already part of mapbox and so no point double bundling. The UMD build still does, since there is no way to access this code in the browser :(

Fixes a bug where arrow-color was accidentally called particle-color. Same for arrow-halo-color.

Moves the demo into docs (so can be served with github pages). In general now the demo is a slideshow of various visualization styles.

Remove various old cruft.

Updates readme with instructions on how to get the data.

The tile json now supports relative URLs.

Fixed a bug that was introduced in some refactoring which was essentially a forgotten this.

Adds CI.